# internal/swagger/swagger.yaml
basePath: /
definitions:
...
  controller.SimpleResponse:
    properties:
      message:
        type: string
      status:
        type: integer
    type: object
  controller.TotpPendingResponse:
    properties:
      message:
        type: string
      status:
        type: integer
      totpPending:
        type: boolean
    type: object
...
info:
  contact: {}
  description: Swagger documentation for Tinyauth's API.
  license:
    name: AGPL-3.0
    url: https://github.com/tinyauthapp/tinyauth/blob/main/LICENSE
  title: Tinyauth API
  version: development
paths:
...
  /api/user/login:
    post:
      consumes:
      - application/json
      description: Login Endpoint
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.TotpPendingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
      summary: Login
      tags:
      - accounts
  /api/user/logout:
    post:
      description: Logout Endpoint
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
      summary: Logout
      tags:
      - accounts
  /api/user/tailscale:
    post:
      consumes:
      - application/json
      description: Tailscale Auth Endpoint (Experimental)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
      summary: Tailscale
      tags:
      - accounts
  /api/user/totp:
    post:
      consumes:
      - application/json
      description: TOTP Endpoint
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
      summary: TOTP
      tags:
      - accounts
...
  /resources/{resource}:
    get:
      description: Get a resource by file name
      parameters:
      - description: Resource Name
        in: path
        name: resource
        required: true
        type: string
      responses:
        "200":
          description: OK
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
      summary: Resources Endpoint
      tags:
      - resources
swagger: "2.0"